feat: refactor schemas into `etherscan/validation`, improve documentation
