+++ src/errors.ts
@@ -65,22 +65,39 @@ export abstract class ParamError extends EtherscanError {}
 
 /* Common Errors */
 
+/**
+ * Indicates that the action call was performed with an invalid API key.
+ */
 export class InvalidApiKeyError extends EtherscanError.extend(
   "Invalid API Key",
 ) {}
 
+/**
+ * Indicates that the action calling frequency surpassed Etherscan rate limits.
+ */
 export class MaxRateLimitError extends EtherscanError.extend(
   "Max rate limit reached, please use API Key for higher rate limit",
 ) {}
 
+/**
+ * Indicates that the action name was not specified during an action call.
+ */
 export class MissingOrInvalidActionError extends EtherscanError.extend(
   "Error! Missing Or invalid Action name",
 ) {}
 
+/**
+ * Indicates that the action call timed out while attempting to respond.
+ */
 export class QueryTimeoutError extends EtherscanError.extend(
   "Query Timeout occured. Please select a smaller result dataset",
 ) {}
 
+/**
+ * Indicates that the error is not implemented in the @railmap/etherscan library.
+ * If you are receiving this error, please
+ * [file an issue](https://github.com/railmap/etherscan/issues/new).
+ */
 export class UnknownError extends EtherscanError.extend(
   "Unknown error. Please file an issue to the @railmap/etherscan repository.",
 ) {}
